requirements.txt
streamlit>=1.34
python-dotenv>=1.0
pypdf>=4.2
faiss-cpu>=1.8
numpy>=1.26
requests>=2.31